+++ lib/rubocop-rspec.rb
@@ -10,8 +10,8 @@ Rubocop::RSpec::Inject.defaults!
 require 'rubocop/cop/top_level_describe'
 
 # cops
-require 'rubocop/cop/rspec_described_class'
-require 'rubocop/cop/rspec_described_method'
-require 'rubocop/cop/rspec_description'
+require 'rubocop/cop/rspec_describe_class'
+require 'rubocop/cop/rspec_describe_method'
+require 'rubocop/cop/rspec_example_wording'
 require 'rubocop/cop/rspec_file_name'
 require 'rubocop/cop/rspec_multiple_describes'