+++ spec/rubocop/cop/rspec_describe_class_spec.rb
@@ -2,7 +2,7 @@
 
 require 'spec_helper'
 
-describe Rubocop::Cop::RSpecDescribedClass do
+describe Rubocop::Cop::RSpecDescribeClass do
   subject(:cop) { described_class.new }
 
   it 'checks first-line describe statements' do